remove typescript 3.8 type im/exports

+++ scripts/types_from_apispec.js
@@ -11,10 +11,34 @@ const http = require('https');
 const fs = require('fs');
 const jsYaml = require('js-yaml');
 const OpenAPI = require('openapi-typescript-codegen');
+const babel = require('@babel/core');
 
 const specFile = __dirname + '/spec.yaml';
 const specUrl = "https://ce85b99cc46752fffee35cab9a7b0278abb4c2d2055cff685af4912c49490f8.gos3.io/spec.yaml";
 const distDir = __dirname + '/../src/Specs';
+
+const removeTypeImExportRecursive = (distDir) => {
+    fs.readdir(distDir, (err, files) => {
+        if (err) {
+            console.error('readdir err', err);
+            return;
+        }
+        files.forEach(file => {
+            const stats = fs.statSync(distDir + '/' + file);
+            if (stats.isDirectory()) {
+                removeTypeImExportRecursive(distDir + '/' + file);
+                return;
+            }
+
+            const source = fs.readFileSync(distDir + '/' + file, 'utf-8');
+            const target = source.replace(/(export|import) type \{/mg, '$1 {');
+            fs.writeFileSync(distDir + '/' + file, target);
+            console.log(distDir + '/' + file + ' - written');
+            
+            
+        });
+    });
+}
 
 // first download the current API Spec
 const file = fs.createWriteStream(specFile);
@@ -41,7 +65,9 @@ http.get(specUrl, function (response) {
             
 
             // clean the dist dir
-            fs.rmSync(distDir, { recursive: true });
+            if (fs.existsSync(distDir)) {
+                fs.rmSync(distDir, { recursive: true });
+            }
 
 
             // ...then generate the types
@@ -52,9 +78,16 @@ http.get(specUrl, function (response) {
                 exportServices: false,
                 input: yaml,
                 output: distDir,
+                useUnionTypes: true
             }).then(() => {
-                console.log('Done!');
+                console.log('Done generating from OpenAPI, now babeling...');
+
+                removeTypeImExportRecursive(distDir);
+
+                
             });
+
+            
             
         });
     });
